+++ app/domain/adminapi/actions.py
@@ -41,9 +41,6 @@ class Actions(adminapi.Actions):
     def create_table(self, r: adminapi.CreateTableRequest) -> tuple[adminapi.CreateTableResponse, bool]:
         return self.table_upload_manager.create_table(r)
 
-    def get_table_validation(self, r: adminapi.GetTableValidationRequest) -> adminapi.GetTableValidationResponse:
-        return self.table_upload_manager.validate_table(r)
-
     def patch_table(self, r: adminapi.PatchTableRequest) -> adminapi.PatchTableResponse:
         return self.table_upload_manager.patch_table(r)
